requirements.txt
streamlit>=1.28.0
Pillow>=9.0.0
opencv-python>=4.6.0
numpy>=1.21.0
pandas>=1.4.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.25.0
scikit-learn>=1.1.0
tqdm>=4.64.0
matplotlib>=3.5.0
easyocr>=1.6.0
